+++ mobile/src/app/(auth)/index.tsx
@@ -1,5 +1,7 @@
 import { useEffect, useState } from 'react';
-import { Alert, View } from 'react-native';
+import { Alert, Text, View } from 'react-native';
+
+import { useSession } from '@/src/app/context/SessionProvider';
 import {
   Card,
   CardContent,
@@ -8,19 +10,37 @@ import {
   CardTitle,
 } from '@/src/components/ui/card';
 import { Checkbox } from '@/src/components/ui/checkbox';
-
 import { supabase } from '@/src/lib/supabase';
-import { useSession } from '@/src/app/context/SessionProvider';
 
 export default function App() {
   const session = useSession();
   const [notes, setNotes] = useState([]);
+  const [name, setName] = useState('');
 
   useEffect(() => {
     if (session) {
       getNotes();
+      getName();
     }
   }, [session]);
+
+  async function getName() {
+    try {
+      const { data } = await supabase
+        .from('profiles')
+        .select(`username, website, avatar_url`)
+        .eq('id', session?.user.id)
+        .single();
+
+      if (data) {
+        setName(data.username);
+      }
+    } catch (error) {
+      if (error instanceof Error) {
+        Alert.alert(error.message);
+      }
+    }
+  }
 
   async function getNotes() {
     try {
@@ -40,26 +60,25 @@ export default function App() {
   }
 
   return (
-    <>
-      <View>
-        {notes?.map((note) => (
-          <Card key={note.id} className='m-4'>
-            <CardHeader>
-              <CardTitle>{note.title}</CardTitle>
-              <CardDescription>
-                {new Date(note.updated_at).toLocaleString('en-US', {
-                  year: 'numeric',
-                  month: 'long',
-                  day: 'numeric',
-                })}
-              </CardDescription>
-            </CardHeader>
-            <CardContent>
-              <Checkbox checked={note.done} onCheckedChange={() => {}} />
-            </CardContent>
-          </Card>
-        ))}
-      </View>
-    </>
+    <View className='p-3'>
+      <Text className='my-4 text-xl font-bold'>👋, {name}</Text>
+      {notes?.map((note) => (
+        <Card key={note.id} className='my-1'>
+          <CardHeader>
+            <CardTitle>{note.title}</CardTitle>
+            <CardDescription>
+              {new Date(note.updated_at).toLocaleString('en-US', {
+                year: 'numeric',
+                month: 'long',
+                day: 'numeric',
+              })}
+            </CardDescription>
+          </CardHeader>
+          <CardContent>
+            <Checkbox checked={note.done} onCheckedChange={() => {}} />
+          </CardContent>
+        </Card>
+      ))}
+    </View>
   );
 }
